backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.8.1
pdfplumber==0.11.5
pypdf==5.4.0
PyMuPDF==1.25.5
pytesseract==0.3.13
rapidfuzz==3.13.0
openai==1.75.0
python-dotenv==1.0.1
reportlab==4.4.1
Pillow==11.2.1
